requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
qrcode==7.4.2
psycopg2-binary==2.9.7
gunicorn==21.2.0